Fix budget groups migration: campaign_id and budget_group_id must be BIGINT
ads_campaigns.id and campaign_budget_groups.id are BIGINT; INT FK columns
cause errno 150.

+++ flaskapp/migrations/create_campaign_budget_groups.sql
@@ -49,8 +49,8 @@ CREATE TABLE IF NOT EXISTS campaign_budget_assignments (
     id INT AUTO_INCREMENT PRIMARY KEY,
 
     -- Campaign and group
-    campaign_id INT NOT NULL,
-    budget_group_id INT NOT NULL,
+    campaign_id BIGINT NOT NULL,
+    budget_group_id BIGINT NOT NULL,
 
     -- Assignment metadata
     assigned_at DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
@@ -74,7 +74,7 @@ CREATE TABLE IF NOT EXISTS campaign_budget_group_spend (
     id INT AUTO_INCREMENT PRIMARY KEY,
 
     -- Group and time period
-    budget_group_id INT NOT NULL,
+    budget_group_id BIGINT NOT NULL,
     period_month DATE NOT NULL,  -- First day of the month (e.g., 2025-01-01)
 
     -- Spend tracking
